+++ packages/@overeng/restate-effect/package.json.genie.ts
@@ -1,4 +1,4 @@
-// @genie-phase bootstrap
+// @genie-bootstrap
 import { otelSdkDeps } from '../../../genie/external.ts'
 import {
   catalog,